requirements.txt
MetaTrader5
python-dotenv
pandas
numpy
ta
requests